Connector failover changes
