+++ src/config/production.json
@@ -16,7 +16,8 @@
     "UserTokenExpirationIntervalSeconds": 3600,
     "FogTokenExpirationIntervalSeconds": 3600,
     "FogStatusUpdateIntervalSeconds": 120,
-    "FogStatusFrequencySeconds": 60
+    "FogStatusFrequencySeconds": 60,
+    "ConnectorHealthCheckFrequencySeconds": 10
   },
   "Database": {
     "Provider": "sqlite",
