+++ src/services/connector-port-service.js
@@ -11,13 +11,11 @@
  *
  */
 
-const ConnectorManager = require('../data/managers/connector-manager')
-const https = require('https')
-const http = require('http')
-const constants = require('../helpers/constants')
-const logger = require('../logger')
 const qs = require('qs')
-const fs = require('fs')
+
+const ConnectorManager = require('../data/managers/connector-manager')
+const { connectorApiHelper } = require('../helpers/http-request-helper')
+const logger = require('../logger')
 
 async function openPortOnRandomConnector (isPublicAccess, transaction) {
   let isConnectorPortOpen = false
@@ -43,40 +41,21 @@ async function openPortOnRandomConnector (isPublicAccess, transaction) {
   return { ports: ports, connector: connector }
 }
 
-async function _openPortsOnConnector (connector, isPublicAccess) {
+function _openPortsOnConnector (connector, isPublicAccess) {
   const data = isPublicAccess
-    ? await qs.stringify({
+    ? qs.stringify({
       mapping: '{"type":"public","maxconnections":60,"heartbeatabsencethreshold":200000}'
     })
-    : await qs.stringify({
+    : qs.stringify({
       mapping: '{"type":"private","maxconnectionsport1":1, "maxconnectionsport2":1, ' +
       '"heartbeatabsencethresholdport1":200000, "heartbeatabsencethresholdport2":200000}'
     })
 
-  const port = connector.devMode ? constants.CONNECTOR_HTTP_PORT : constants.CONNECTOR_HTTPS_PORT
-
-  const options = {
-    host: connector.domain,
-    port: port,
-    path: '/api/v2/mapping/add',
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/x-www-form-urlencoded',
-      'Content-Length': Buffer.byteLength(data)
-    }
-  }
-  if (!connector.devMode && connector.cert && connector.isSelfSignedCert === true) {
-    const ca = fs.readFileSync(connector.cert)
-    /* eslint-disable new-cap */
-    options.ca = new Buffer.from(ca)
-  }
-
-  const ports = await _makeRequest(connector, options, data)
-  return ports
+  return connectorApiHelper(connector, data, '/api/v2/mapping/add')
 }
 
 async function _getRandomConnector (transaction) {
-  const connectors = await ConnectorManager.findAll({}, transaction)
+  const connectors = await ConnectorManager.findAll({ healthy: true }, transaction)
 
   if (connectors && connectors.length > 0) {
     const randomNumber = Math.round((Math.random() * (connectors.length - 1)))
@@ -86,63 +65,12 @@ async function _getRandomConnector (transaction) {
   }
 }
 
-async function closePortOnConnector (connector, ports) {
+function closePortOnConnector (connector, ports) {
   const data = qs.stringify({
     mappingid: ports.mappingId
   })
 
-  const port = connector.devMode ? constants.CONNECTOR_HTTP_PORT : constants.CONNECTOR_HTTPS_PORT
-
-  const options = {
-    host: connector.domain,
-    port: port,
-    path: '/api/v2/mapping/remove',
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/x-www-form-urlencoded',
-      'Content-Length': Buffer.byteLength(data)
-    }
-  }
-  if (!connector.devMode && connector.cert && connector.isSelfSignedCert === true) {
-    const ca = fs.readFileSync(connector.cert)
-    /* eslint-disable new-cap */
-    options.ca = new Buffer.from(ca)
-  }
-
-  await _makeRequest(connector, options, data)
-}
-
-async function _makeRequest (connector, options, data) {
-  return new Promise((resolve, reject) => {
-    const httpreq = (connector.devMode ? http : https).request(options, function (response) {
-      let output = ''
-      response.setEncoding('utf8')
-
-      response.on('data', function (chunk) {
-        output += chunk
-      })
-
-      response.on('end', function () {
-        const responseObj = JSON.parse(output)
-        if (responseObj.errormessage) {
-          return reject(new Error(responseObj.errormessage))
-        } else {
-          return resolve(responseObj)
-        }
-      })
-    })
-
-    httpreq.on('error', function (err) {
-      if (err instanceof Error) {
-        return reject(new Error(err.message))
-      } else {
-        return reject(new Error(JSON.stringify(err)))
-      }
-    })
-
-    httpreq.write(data)
-    httpreq.end()
-  })
+  return connectorApiHelper(connector, data, '/api/v2/mapping/remove')
 }
 
 module.exports = {
